test: WOW-1B CLI 테스트 support 추출
